Refactor Hero and Enemy classes. Place general logic to Actor class
