+++ core/src/main/java/com/cursedarchie/platformer/actors/enemies/DefaultEnemy.java
@@ -1,6 +1,7 @@
 package com.cursedarchie.platformer.actors.enemies;
 
 import com.badlogic.gdx.math.Vector2;
+import com.cursedarchie.platformer.actors.Enemy;
 
 public class DefaultEnemy extends Enemy {
