--- core/src/main/java/com/cursedarchie/platformer/actors/enemies/Enemy.java
@@ -1,268 +1,0 @@
-package com.cursedarchie.platformer.actors.enemies;
-
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.math.Rectangle;
-import com.badlogic.gdx.math.Vector2;
-
-public abstract class Enemy {
-
-    public enum EnemyState {
-        IDLE, WALKING, JUMPING, ATTACKING, DYING
-    }
-
-    public float SIZE;
-    private float MAX_HEALTH;
-    private float MAX_ACCELERATION;
-    private float health;
-    private float damage;
-    private float ATTACK_DURATION;
-    private float VIEW_DISTANCE;
-    private float VIEW_ANGLE;
-
-    Vector2 position = new Vector2();
-    Vector2 acceleration = new Vector2();
-    Vector2 velocity = new Vector2();
-    Rectangle bounds = new Rectangle();
-    private EnemyState enemyState = EnemyState.IDLE;
-    private boolean facingLeft = true;
-    private float stateTime = 0f;
-    private boolean grounded = false;
-    private boolean isAlive = true;
-    private boolean damageDealt = false;
-    private boolean isHeroInSight = false;
-
-    private float attackTime = 0f;
-
-    public Enemy(Vector2 pos) {
-        this.position.set(pos);
-        this.bounds.x = pos.x;
-        this.bounds.y = pos.y;
-        this.bounds.width = SIZE;
-        this.bounds.height = SIZE;
-    }
-
-    /**
-     *  @return This method return true when enemy is facing left
-     */
-    public boolean isFacingLeft() {
-        return facingLeft;
-    }
-    public void setFacingLeft(boolean facingLeft) {
-        this.facingLeft = facingLeft;
-    }
-
-    /**
-     * @return This method return true when enemy stay on any block
-     */
-    public boolean isGrounded() {
-        return grounded;
-    }
-    public void setGrounded(boolean grounded) {
-        this.grounded = grounded;
-    }
-
-    /**
-     * @return This method return position of enemy coordinates (x, y)
-     */
-    public Vector2 getPosition() {
-        return position;
-    }
-    public void setPosition(Vector2 position) {
-        this.position.set(position);
-        this.bounds.setX(position.x);
-        this.bounds.setY(position.y);
-    }
-
-    /**
-     * @return This method return acceleration limits of the enemy
-     */
-    public float getMaxAcceleration() {
-        return MAX_ACCELERATION;
-    }
-    public void setMaxAcceleration(float maxAcceleration) {
-        this.MAX_ACCELERATION = maxAcceleration;
-    }
-
-    /**
-     * @return current acceleration
-     */
-    public Vector2 getAcceleration() {
-        return acceleration;
-    }
-    public void setAcceleration(Vector2 acceleration) {
-        this.acceleration.set(acceleration);
-    }
-
-    /**
-     * @return current speed of enemy
-     */
-    public Vector2 getVelocity() {
-        return velocity;
-    }
-    public void setVelocity(Vector2 velocity) {
-        this.velocity.set(velocity);
-    }
-
-    /**
-     * @return square where enemy is located (x, y, width, height)
-     */
-    public Rectangle getBounds() {
-        return bounds;
-    }
-    public void setBounds(Rectangle bounds) {
-        this.bounds.set(bounds);
-    }
-
-    /**
-     * @return time which enemy staying in current EnemyState
-     */
-    public float getStateTime() {
-        return stateTime;
-    }
-    public void setStateTime(float stateTime) {
-        this.stateTime = stateTime;
-    }
-
-    /**
-     * @return current EnemyState
-     */
-    public EnemyState getState() {
-        return enemyState;
-    }
-    public void setState(EnemyState enemyState) {
-        if (this.enemyState != enemyState) {
-            Gdx.app.log("Enemy", "EnemyState change: " + this.enemyState + " -> " + enemyState);
-            this.enemyState = enemyState;
-            stateTime = 0f;
-        }
-
-    }
-
-    /**
-     * @return SIZE. this.bounds.width = this.bounds.height = SIZE
-     */
-    public float getSize() {
-        return SIZE;
-    }
-    public void setSize(float SIZE) {
-        this.SIZE = SIZE;
-    }
-
-    /**
-     * @return True when enemy can see hero
-     */
-    public boolean isHeroInSight() {
-        return isHeroInSight;
-    }
-    public void setHeroInSight(boolean heroInSight) {
-        isHeroInSight = heroInSight;
-    }
-
-    /**
-     * @return float of Angle of the field of view
-     */
-    public float getViewAngle() {
-        return VIEW_ANGLE;
-    }
-    public void setViewAngle(float VIEW_ANGLE) {
-        this.VIEW_ANGLE = VIEW_ANGLE;
-    }
-
-    /**
-     * @return the distance of view
-     */
-    public float getViewDistance() {
-        return VIEW_DISTANCE;
-    }
-    public void setViewDistance(float VIEW_DISTANCE) {
-        this.VIEW_DISTANCE = VIEW_DISTANCE;
-    }
-
-    /**
-     * @return the time before causing damage
-     */
-    public float getAttackDuration() {
-        return ATTACK_DURATION;
-    }
-    public void setAttackDuration(float ATTACK_DURATION) {
-        this.ATTACK_DURATION = ATTACK_DURATION;
-    }
-
-    /**
-     * @param damage the damage is will take
-     */
-    public void takeDamage (float damage) {
-        this.health -= damage;
-        Gdx.app.log("INFO", "TAKE DAMAGE. HP: " + this.getHealth());
-    }
-
-    /**
-     * @return the enemy's damage
-     */
-    public float getDamage() {
-        return damage;
-    }
-    public void setDamage(float damage) {
-        this.damage = damage;
-    }
-
-    /**
-     * @Description: checks if already attacking. If not, set state and start count time before damage dealing.
-     */
-    public void startAttack() {
-        if (getState() != EnemyState.ATTACKING) {
-            setState(EnemyState.ATTACKING);
-            attackTime = 0f;
-            damageDealt = false;
-        }
-    }
-
-    /**
-     * @param delta unit of time
-     * @Description: if enemy attacks update time before damage dealing, if enemy had already attacked set IDLE state
-     */
-    public void updateAttack(float delta) {
-        if (getState() == EnemyState.ATTACKING) {
-            attackTime += delta;
-            if (attackTime >= ATTACK_DURATION) {
-                setState(EnemyState.IDLE);
-            }
-        }
-    }
-
-    /**
-     * @return true if hero so nearby to deal damage
-     */
-    public boolean canDealDamage() {
-        return getState() == EnemyState.ATTACKING && !damageDealt && attackTime >= ATTACK_DURATION / 2f;
-    }
-    public void markDamageDealt() {
-        damageDealt = true;
-    }
-
-    public float getMaxHealth() {
-        return MAX_HEALTH;
-    }
-    public void setMaxHealth(float maxHealth) {
-        MAX_HEALTH = maxHealth;
-    }
-    public float getHealth() {
-        return health;
-    }
-    public void setHealth(float health) {
-        this.health = health;
-    }
-    public boolean isAlive() {
-        return isAlive;
-    }
-    public void setAlive(boolean alive) {
-        isAlive = alive;
-    }
-
-    public void update(float delta) {
-        position.add(velocity.cpy().scl(delta));
-        bounds.x = position.x;
-        bounds.y = position.y;
-        stateTime += delta;
-    }
-}
